+++ src/filesprocessing/filters/SmallerThanFilter.java
@@ -3,7 +3,7 @@ package filesprocessing.filters;
 import java.io.File;
 
 public class SmallerThanFilter extends Filter{
-	private static final int SIZE_FACTOR = 1024;
+	private static final long SIZE_FACTOR = 1024;
 	public boolean passFilter(File f, String[] args) throws FilterException {
 		long bound =  (long) Double.parseDouble(args[1].replace(" ",""));
